requirements.txt
numpy<2.0.0
rembg
opencv-contrib-python-headless
Pillow>=9.0.0
streamlit>=1.28.0
onnxruntime
pymatting
torch>=1.13.0,<2.1.0
torchvision>=0.14.0,<0.16.0
scikit-image>=0.19.0,<0.21.0
scipy>=1.9.0,<1.11.0
scikit-learn>=1.1.0,<1.3.0
matplotlib>=3.5.0
tqdm>=4.64.0